requirements.txt
requests
pycryptodome
requests_toolbelt